+++ src/AIDevelopers/Interfaces/DamageInterface.java
@@ -3,5 +3,5 @@ package AIDevelopers.Interfaces;
 import AIDevelopers.Entities.Andrey;
 
 public interface DamageInterface {
-    public void ApplyDamage(Andrey damagedAndrey, int DamageToApply);
+    public void ApplyDamage(Andrey damagedAndrey, float DamageToApply);
 }
